Separate InteractiveMenu impl to `basic` version
