+++ src/main/java/com/freya02/botcommands/api/pagination/interactive/BasicInteractiveMenuBuilder.java
@@ -1,0 +1,33 @@
+package com.freya02.botcommands.api.pagination.interactive;
+
+import com.freya02.botcommands.api.pagination.BasicPaginationBuilder;
+import net.dv8tion.jda.api.entities.Emoji;
+import net.dv8tion.jda.api.interactions.components.selections.SelectOption;
+import org.jetbrains.annotations.NotNull;
+
+import java.util.ArrayList;
+import java.util.List;
+
+/**
+ * @param <T> Type of the implementor
+ * @param <R> Type of the implementor {@link #build()} return type
+ */
+@SuppressWarnings("unchecked")
+public abstract class BasicInteractiveMenuBuilder<T extends BasicInteractiveMenuBuilder<T, R>, R extends BasicInteractiveMenu<R>> extends BasicPaginationBuilder<T, R> {
+	protected final List<InteractiveMenuItem> items = new ArrayList<>();
+
+	/**
+	 * Adds a menu to this {@link InteractiveMenu}
+	 * <br><b>Note: The first added menu will be the first selected one</b>
+	 *
+	 * @param content  The content of the {@link SelectOption} bound to this menu
+	 * @param supplier The interactive menu supplier for this menu's page
+	 * @return This builder for chaining convenience
+	 * @see SelectContent#of(String, String, Emoji)
+	 */
+	public T addMenu(@NotNull SelectContent content, @NotNull InteractiveMenuSupplier supplier) {
+		items.add(new InteractiveMenuItem(content, supplier));
+
+		return (T) this;
+	}
+}
