+++ src/main/java/com/freya02/botcommands/api/pagination/interactive/InteractiveMenu.java
@@ -1,15 +1,8 @@
 package com.freya02.botcommands.api.pagination.interactive;
 
-import com.freya02.botcommands.api.components.Components;
 import com.freya02.botcommands.api.components.InteractionConstraints;
-import com.freya02.botcommands.api.components.builder.LambdaSelectionMenuBuilder;
-import com.freya02.botcommands.api.components.event.SelectionEvent;
-import com.freya02.botcommands.api.pagination.BasicPagination;
 import com.freya02.botcommands.api.pagination.TimeoutInfo;
-import net.dv8tion.jda.api.entities.Message;
-import net.dv8tion.jda.api.entities.MessageEmbed;
 import net.dv8tion.jda.api.interactions.components.selections.SelectMenu;
-import net.dv8tion.jda.api.interactions.components.selections.SelectOption;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
 
@@ -17,55 +10,11 @@ import java.util.List;
 
 /**
  * A type of pagination which shows embeds and provides a {@link SelectMenu} to navigate between menus
+ * <br>Each embed is bound to a selection menu
+ * <br><i>This does not provide pagination for each embed</i> (no arrow buttons, only the selection menu)
  */
-public class InteractiveMenu extends BasicPagination<InteractiveMenu> {
-	private final List<InteractiveMenuItem> items;
-
-	private int selectedItem = 0;
-
+public final class InteractiveMenu extends BasicInteractiveMenu<InteractiveMenu> {
 	InteractiveMenu(@NotNull List<InteractiveMenuItem> items, InteractionConstraints constraints, @Nullable TimeoutInfo<InteractiveMenu> timeout) {
-		super(constraints, timeout);
-
-		if (items.isEmpty()) throw new IllegalStateException("No interactive menu items has been added");
-
-		this.items = items;
-	}
-
-	@NotNull
-	private SelectMenu buildSelectMenu() {
-		final LambdaSelectionMenuBuilder builder = Components.selectionMenu(this::handleSelection).oneUse().setConstraints(constraints);
-
-		final List<SelectOption> options = builder.getOptions();
-		for (int i = 0, itemsSize = items.size(); i < itemsSize; i++) {
-			InteractiveMenuItem item = items.get(i);
-
-			SelectOption option = item.content().toSelectOption(String.valueOf(i));
-			if (i == selectedItem) option = option.withDefault(true);
-
-			options.add(option);
-		}
-
-		return builder.build();
-	}
-
-	private void handleSelection(SelectionEvent event) {
-		selectedItem = Integer.parseInt(event.getValues().get(0));
-
-		event.editMessage(get()).queue();
-	}
-
-	@Override
-	public Message get() {
-		onPreGet();
-
-		components.addComponents(0, buildSelectMenu());
-
-		final MessageEmbed embed = items.get(selectedItem).supplier().get(messageBuilder, components);
-		messageBuilder.setEmbeds(embed);
-		messageBuilder.setActionRows(components.getActionRows());
-
-		onPostGet();
-
-		return messageBuilder.build();
+		super(items, constraints, timeout);
 	}
 }
