+++ src/main/java/com/freya02/botcommands/api/pagination/interactive/InteractiveMenuBuilder.java
@@ -1,31 +1,11 @@
 package com.freya02.botcommands.api.pagination.interactive;
 
-import com.freya02.botcommands.api.pagination.BasicPaginationBuilder;
-import net.dv8tion.jda.api.entities.Emoji;
-import net.dv8tion.jda.api.interactions.components.selections.SelectOption;
 import org.jetbrains.annotations.NotNull;
 
-import java.util.ArrayList;
-import java.util.List;
-
-public class InteractiveMenuBuilder extends BasicPaginationBuilder<InteractiveMenuBuilder, InteractiveMenu> {
-	private final List<InteractiveMenuItem> items = new ArrayList<>();
-
-	/**
-	 * Adds a menu to this {@link InteractiveMenu}
-	 * <br><b>Note: The first added menu will be the first selected one</b>
-	 *
-	 * @param content  The content of the {@link SelectOption} bound to this menu
-	 * @param supplier The interactive menu supplier for this menu's page
-	 * @return This builder for chaining convenience
-	 * @see SelectContent#of(String, String, Emoji)
-	 */
-	public InteractiveMenuBuilder addMenu(@NotNull SelectContent content, @NotNull InteractiveMenuSupplier supplier) {
-		items.add(new InteractiveMenuItem(content, supplier));
-
-		return this;
-	}
-
+/**
+ * Builds an {@link InteractiveMenu}
+ */
+public final class InteractiveMenuBuilder extends BasicInteractiveMenuBuilder<InteractiveMenuBuilder, InteractiveMenu> {
 	@Override
 	@NotNull
 	public InteractiveMenu build() {
